DCMAW-8229: rename mock jwt builder file
